Move statsd_informant.rb to statsd/informant.rb
